Update NetwrixFileShareSetup.ps1
Corrected a fundamental misunderstanding - the user in the parameter for System.Security.AccessControl.FileSystemAuditRule is not who has access to the logs, but instead which user accounts to record the actions of. This means Everyone is the most common choice, unless you're only auditing a few users. Left in there for configuration potential but not requesting data with Read-Host still as it's uncommon to change that value.